# src/app/api/auth/route.ts
import { NextResponse } from "next/server";
import bcrypt from "bcryptjs";
import { z } from "zod";
import { getCollection } from "@/lib/db";
import { issueToken } from "@/lib/auth";

const registerSchema = z.object({
  email: z.string().email(),
  password: z.string().min(6),
  name: z.string().optional(),
});

const loginSchema = z.object({
  email: z.string().email(),
  password: z.string().min(6),
});

export async function POST(req: Request) {
  const json = await req.json().catch(() => null);
  const action = json?.action as "register" | "login" | undefined;

  if (action === "register") {
    const parsed = registerSchema.safeParse(json);
    if (!parsed.success) {
      return NextResponse.json({ error: "Invalid input", details: parsed.error.flatten() }, { status: 400 });
    }

    const { email, password, name } = parsed.data;
    const users = await getCollection("users");

    const existing = await users.findOne({ email });
    if (existing) {
      return NextResponse.json({ error: "User already exists" }, { status: 409 });
    }

    const passwordHash = await bcrypt.hash(password, 10);
    const user = {
      email,
      passwordHash,
      name,
      createdAt: new Date(),
    };

    const result = await users.insertOne(user);
    const token = issueToken(result.insertedId.toString());
    return NextResponse.json({ token, user: { id: result.insertedId, email, name } }, { status: 201 });
  }

  if (action === "login") {
    const parsed = loginSchema.safeParse(json);
    if (!parsed.success) {
      return NextResponse.json({ error: "Invalid input", details: parsed.error.flatten() }, { status: 400 });
    }

    const { email, password } = parsed.data;
    const users = await getCollection("users");
    const user = await users.findOne({ email });
    if (!user) {
      return NextResponse.json({ error: "Invalid credentials" }, { status: 401 });
    }

    const ok = await bcrypt.compare(password, (user as any).passwordHash);
    if (!ok) {
      return NextResponse.json({ error: "Invalid credentials" }, { status: 401 });
    }

    const token = issueToken((user as any)._id.toString());
    return NextResponse.json({ token, user: { id: (user as any)._id, email: user.email, name: user.name } });
  }

  return NextResponse.json({ error: "Unknown action. Use 'register' or 'login'" }, { status: 400 });
}